Premiestnenie utility funkcii do oddeleneho suboru
